Move categorization to SCA service. It is called before the number of testCaseFeedbacks is checked. Otherwise SCA feedback won't get categorized when the build fails.
